Create modular (and eventually portable) workspace.
