+++ resolve/LinSolverDirectCuSolverGLU.hpp
@@ -1,6 +1,5 @@
 #pragma once
 #include "Common.hpp"
-#include <resolve/LinAlgWorkspace.hpp>
 #include "LinSolver.hpp"
 #include "cusolver_defs.hpp"
 #include <resolve/MemoryUtils.hpp>
@@ -18,6 +17,9 @@ namespace ReSolve
   {
     class Sparse;
   }
+
+  // Forward declaration of ReSolve handlers workspace
+  class LinAlgWorkspace;
 
   class LinSolverDirectCuSolverGLU : public LinSolverDirect 
   {
@@ -38,7 +40,7 @@ namespace ReSolve
       //note: we need cuSolver handle, we can copy it from the workspace to avoid double allocation
       cusparseMatDescr_t descr_M_; //this is NOT sparse matrix descriptor
       cusparseMatDescr_t descr_A_; //this is NOT sparse matrix descriptor
-      LinAlgWorkspace *workspace_;// so we can copy cusparse handle
+      LinAlgWorkspace* workspace_;// so we can copy cusparse handle
       cusolverSpHandle_t handle_cusolversp_; 
       cusolverStatus_t status_cusolver_;
       cusparseStatus_t status_cusparse_;
