+++ resolve/workspace/LinAlgWorkspace.cpp
@@ -1,0 +1,13 @@
+#include "LinAlgWorkspace.hpp"
+
+namespace ReSolve
+{
+  LinAlgWorkspace::LinAlgWorkspace()
+  {
+  }
+  
+  LinAlgWorkspace::~LinAlgWorkspace()
+  {
+  }
+
+}
