+++ resolve/workspace/LinAlgWorkspace.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+
+#include <resolve/MemoryUtils.hpp>
+
+namespace ReSolve
+{
+  class LinAlgWorkspace
+  {
+    public:
+      LinAlgWorkspace();
+      ~LinAlgWorkspace();
+    protected:
+      MemoryHandler mem_;
+  };
+
+}
